Rename Ansible Project vue file
